requirements.txt
# AI Documentation Assistant - Dependencies (FREE VERSION)

# Web Framework
streamlit==1.39.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2

# Embeddings & Vector Search
sentence-transformers==3.3.1
faiss-cpu==1.13.0

# Local AI Models (FREE!)
transformers==4.46.3
torch>=2.6.0

# Utilities
python-dotenv==1.0.1